refactor: Global CSS styles, event page fixes and favicon issues
