--- .vscode/settings.json
@@ -1,12 +1,0 @@
-{
-	"emeraldwalk.runonsave": {
-		"commands": [
-			{
-				// If you have your CSS file somewhere else, you have to specify the
-				// correct path here.
-				"match": "./app/globals-edit.css",
-				"cmd": "npm run tw:fix"
-			}
-		]
-	}
-}
