+++ app/dashboard/events/page.tsx
@@ -10,141 +10,14 @@ import ErrorFaceIcon from "@/public/icons/error_face.svg";
 import SearchBar from "@/components/dashboard/SearchBar";
 import SortBy from "@/components/dashboard/SortBy";
 
-import { AceLabel } from "@/components/dashboard/AceCard";
+import { EventPreview } from "@/components/events/EventPreview";
 import { DashboardPagination } from "@/components/dashboard/Pagination";
 import { Filter } from "@/components/dashboard/Filter";
 import { ParamsResponsiblePicker } from "@/components/ResponsiblePicker";
-import { DateDisplay } from "@/components/ui/calendar";
 
 // Types
-import { z } from "zod";
-import { ACEs, addEventFormSchema } from "@/lib/validations/AddEventForm";
-import { cn } from "@/lib/utils";
-
-type Event = z.infer<typeof addEventFormSchema> & { id: string };
-
-const events: Event[] = [
-	{
-		id: "EHD723",
-		name: "Reunião de planejamento de eventos",
-		description: undefined,
-		responsible: ["1", "2"],
-		dateFrom: new Date("2022-09-15T00:00:00Z"),
-		timeFrom: "14:00",
-		timeTo: "16:00",
-		ace: "3",
-	},
-	{
-		id: "DHD723",
-		name: "Reunião Semanal",
-		description: "Reunião semanal para discutir os eventos da semana",
-		responsible: ["1"],
-		dateFrom: new Date("2022-02-24T00:00:00Z"),
-		timeFrom: "14:00",
-		timeTo: "16:00",
-		ace: "1",
-	},
-	{
-		id: "UHD723",
-		name: "Treino de táticas retiradas de livros",
-		description:
-			"Consiste na realização de sessões de treinamento focadas em aprender táticas retiradas de livros de xadrez relevantes. Os membros estudarão e resolverão problemas com o objetivo de melhorar sua capacidade de visualização e cálculo de jogo.",
-		responsible: ["1", "2"],
-		dateFrom: new Date("2022-09-15T00:00:00Z"),
-		timeFrom: "14:00",
-		timeTo: "16:00",
-		ace: "3",
-	},
-	{
-		id: "IHD723",
-		name: "Análise de partidas históricas",
-		description:
-			"Envolve a análise de partidas históricas de xadrez, permitindo aos membros do grupo aprenderem com partidas passadas que tiveram momentos icônicos.",
-		responsible: ["1", "2"],
-		dateFrom: new Date("2022-09-15T00:00:00Z"),
-		timeFrom: "14:00",
-		timeTo: "16:00",
-		ace: "3",
-	},
-	{
-		id: "OHD723",
-		name: "Análise de partidas relevantes entre os membros",
-		description:
-			"Consiste na análise conjunta de partidas realizadas entre os membros da ACE. Dessa forma será possível ver as diferentes abordagens individuais dos jogadores, erros que comeram e pontos a serem fortalecidos.",
-		responsible: ["1", "2"],
-		dateFrom: new Date("2022-09-15T00:00:00Z"),
-		timeFrom: "14:00",
-		timeTo: "16:00",
-		ace: "3",
-	},
-];
-
-const PERIODS = [
-	{
-		id: "2022.1",
-		from: new Date("2022-01-01T00:00:00Z"),
-		to: new Date("2022-06-30T23:59:59Z"),
-	},
-	{
-		id: "2022.2",
-		from: new Date("2022-07-01T00:00:00Z"),
-		to: new Date("2022-12-31T23:59:59Z"),
-	},
-	{
-		id: "2023.1",
-		from: new Date("2023-01-01T00:00:00Z"),
-		to: new Date("2023-06-30T23:59:59Z"),
-	},
-	{
-		id: "2023.2",
-		from: new Date("2023-07-01T00:00:00Z"),
-		to: new Date("2023-12-31T23:59:59Z"),
-	},
-	{
-		id: "2024.1",
-		from: new Date("2024-01-01T00:00:00Z"),
-		to: new Date("2024-06-30T23:59:59Z"),
-	},
-	{
-		id: "2024.2",
-		from: new Date("2024-07-01T00:00:00Z"),
-		to: new Date("2024-12-31T23:59:59Z"),
-	},
-];
-
-interface Moderator {
-	name: string;
-	image_url: string;
-	id: string;
-}
-
-const moderators = [
-	{
-		name: "John Doe",
-		image_url: "https://randomuser.me/api/portraits/men/18.jpg",
-		id: "1",
-	},
-	{
-		name: "Marcelo Silva",
-		image_url: "https://randomuser.me/api/portraits/men/19.jpg",
-		id: "2",
-	},
-	{
-		name: "Marcos Melo",
-		image_url: "https://randomuser.me/api/portraits/men/20.jpg",
-		id: "3",
-	},
-	{
-		name: "Luciano Cesa",
-		image_url: "https://github.com/mark.png",
-		id: "4",
-	},
-	{
-		name: "Márcio Cavalcante",
-		image_url: "https://randomuser.me/api/portraits/men/21.jpg",
-		id: "5",
-	},
-] as Moderator[];
+import { ACEs } from "@/lib/validations/AddEventForm";
+import { PERIODS, events } from "@/lib/fake_data";
 
 const EVENTS_PER_PAGE = 3;
 
@@ -318,84 +191,6 @@ function Empty() {
 			>
 				Limpar filtros
 			</Link>
-		</div>
-	);
-}
-
-function EventPreview({ event }: { event: Event }) {
-	const ACE = ACEs.find((ace) => ace.id == event.ace);
-
-	const responsible = event.responsible.map((id) => {
-		const mod = moderators.find((mod) => mod.id === id);
-		return mod;
-	});
-	const lastResponsible = responsible.length > 1 ? responsible.pop() : null;
-
-	return (
-		<li className="flex flex-col items-start justify-start p-9 gap-4 bg-gray-300 rounded-lg w-full hover:bg-gray-400 hover:outline outline-gray-200 transition-[background-color,outline]">
-			<div className="flex flex-row items-center justify-between flex-wrap w-full gap-2">
-				<h3 className="text-lg lg:text-xl font-extrabold font-title leading-snug text-left">
-					{event.name}
-				</h3>
-				<p className="opacity-50 text-neutral text-xs lg:text-base font-semibold leading-none">
-					#{event.id}
-				</p>
-			</div>
-			{event.description && (
-				<p className="text-sm lg:text-base text-muted font-medium">
-					{event.description}
-				</p>
-			)}
-			<div className="flex flex-row flex-wrap items-center gap-4 justify-between w-full">
-				{ACE && <AceLabel ace={ACE} />}
-				<div className="flex flex-row items-center justify-between">
-					<DateDisplay
-						dateString={event.dateFrom.toLocaleDateString("pt-BR", {
-							year: "numeric",
-							month: "2-digit",
-							day: "numeric",
-						})}
-					/>
-					{/* <span className="text-sm font-medium">{event.timeFrom}</span> */}
-				</div>
-			</div>
-			<div className="flex flex-row flex-wrap items-center justify-between pt-4 border-t border-t-gray-100 w-full gap-4">
-				<div className="flex flex-row items-center justify-start gap-2 max-sm:w-full">
-					<ProfileImages
-						image_urls={responsible
-							.map((mod) => mod!.image_url)
-							.concat(lastResponsible?.image_url || [])}
-					/>
-					<span className="text-left text-neutral text-sm font-semibold">
-						Organizado por{" "}
-						{responsible.map((mod) => mod?.name).join(", ")}{" "}
-						{lastResponsible && "e"} {lastResponsible?.name}
-					</span>
-				</div>
-				<p className="text-right text-neutral text-sm font-semibold">
-					+ de <span className="underline">10 membros</span>{" "}
-					participaram
-				</p>
-			</div>
-		</li>
-	);
-}
-
-function ProfileImages({ image_urls }: { image_urls: string[] }) {
-	return (
-		<div className="flex flex-row items-center justify-start">
-			{image_urls.map((url) => (
-				<Image
-					src={url}
-					alt="Profile Image"
-					key={url}
-					height={24}
-					width={24}
-					className={cn("w-6 h-6 min-w-6 rounded-full", {
-						"-ml-2": image_urls.indexOf(url) > 0,
-					})}
-				/>
-			))}
 		</div>
 	);
 }
