--- app/globals-edit.css
@@ -1,104 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-	.font-outline-1 {
-		-webkit-text-stroke: 1px hsl(var(--neutral) / 0.1);
-	}
-	.font-outline-2 {
-		-webkit-text-stroke: 1px hsl(var(--neutral) / 0.05);
-	}
-
-	:root {
-		--neutral: hsl(0 0% 0%);
-		--foreground: hsl(0 0% 30.49%);
-
-		--primary-100: hsl(90 39% 51%);
-		--primary-200: hsl(90 30% 42%);
-
-		--secondary-100: hsl(38 79% 58%);
-		--secondary-200: hsl(38 68% 49%);
-
-		--tertiary-100: hsl(19 58% 52%);
-		--tertiary-200: hsl(19 64% 43%);
-
-		--info-100: hsl(216 58% 52%);
-		--info-200: hsl(216 67% 44%);
-
-		--gray-200: hsl(240 100% 98.43%);
-		--gray-300: hsl(20 14.29% 95.88%);
-		--gray-400: hsl(0 0% 88.24%);
-		--gray-500: hsl(0 0% 80.39%);
-		--gray-600: hsl(0 0% 83.39%);
-
-		--wrapper: 1.44rem;
-		--radius: 0.5rem;
-	}
-
-	.dark {
-		--neutral: hsl(0 0% 100%);
-		--foreground: hsl(0 0% 85.49%);
-
-		--gray-100: 35 8.96 26.27%;
-		--gray-200: hsl(40 12% 20%);
-		--gray-300: hsl(38 12% 18%);
-		--gray-400: hsl(40 8% 15%);
-		--gray-500: hsl(30 3% 11%);
-		--gray-600: hsl(30 4% 9%);
-	}
-}
-
-html {
-	scroll-behavior: smooth;
-}
-
-input[type="number"]::-webkit-inner-spin-button,
-input[type="number"]::-webkit-outer-spin-button {
-	-webkit-appearance: none;
-	margin: 0;
-}
-
-.no-scrollbar {
-	-ms-overflow-style: none;
-	scrollbar-width: none;
-	&::-webkit-scrollbar {
-		display: none;
-	}
-}
-
-.thin-scrollbar {
-	&::-webkit-scrollbar {
-		width: 0.5rem;
-	}
-	&::-webkit-scrollbar-track {
-		background-color: hsl(var(--gray-200));
-		border-radius: 0.25rem;
-	}
-	&::-webkit-scrollbar-thumb {
-		background-color: hsl(var(--gray-400));
-		border-radius: 0.25rem;
-	}
-}
-
-@media (min-width: 640px) {
-	:root {
-		--wrapper: 2.88rem;
-	}
-}
-
-@media (min-width: 1024px) {
-	:root {
-		--wrapper: 7.68rem;
-	}
-}
-
-@layer base {
-	* {
-		@apply border-border;
-	}
-	body {
-		@apply bg-gray-600 text-neutral text-base;
-		font-feature-settings: "rlig" 1, "calt" 1;
-	}
-}
