+++ app/layout.tsx
@@ -36,9 +36,9 @@ export default function RootLayout({
 		<html lang="en" suppressHydrationWarning>
 			<body
 				className={cn(
-					"min-h-screen bg-gray-600 font-sans antialiased relative",
+					"min-h-screen font-sans antialiased relative",
 					fontSans.variable,
-					titleFont.variable
+					titleFont.variable,
 				)}
 			>
 				<ThemeProvider
