--- components/events/Event.model.ts
@@ -1,8 +1,0 @@
-
-export interface EventProps {
-	title?: string;
-	description?: string;
-	date?: string;
-	url?: string;
-	state: "error" | "not-logged" | "join" | "joining" | "read-only"; //ajeitar isso
-}
