+++ package.json
@@ -7,8 +7,7 @@
 		"build": "next build",
 		"start": "next start",
 		"lint": "next lint",
-		"prettier": "bunx prettier . --write",
-		"tw:fix": "node fix-tailwind-vars"
+		"prettier": "pnpm dlx prettier . --write"
 	},
 	"dependencies": {
 		"@designbycode/tailwindcss-text-stroke": "^1.2.1",
